feat: guide field styling tweaks and toggle
